chat/pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = "Messaging API for user interractions via text/IM"
authors = ["raokrutarth"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
uvicorn = "^0.12.2"
fastapi = "^0.61.2"
slack_sdk = "^3.0.0-rc.3"
aiohttp = "^3.7.2"
gunicorn = "^20.0.4"
uvloop = "^0.14.0"
httptools = "^0.1.1"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
mypy = "^0.790"
watchdog = {extras = ["watchmedo"], version = "^0.10.3"}
autoflake = "^1.4"
isort = "^5.6.4"
pylint = "^2.6.0"
PyFlakes = "^2.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
